DevProject_AI/backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
python-dotenv==1.2.2
anthropic==0.86.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
